refactor(task-graph): extract EdgeMaterializer from TaskGraphRunner
Moves dataflow read/write methods (copyInputFromEdgesToNode,
pushOutputFromNodeToEdges, pushErrorFromNodeToEdges, hasErrorOutputEdges,
pushErrorOutputToEdges, filterInputForTask, resetTask) plus the duck-typed
isImageValueShape predicate into a new internal class. The facade owns one
instance, constructed once.

EdgeMaterializer takes the graph at construction so methods don't need a graph
arg — the back-reference matches what the original code already assumed.

Pure refactor: no behavior change, no public API change. 1,663 integration
tests pass unchanged.
